+++ com.unity.cinemachine/Runtime/Core/CinemachinePathBase.cs
@@ -110,13 +110,16 @@ namespace Cinemachine
             float end = MaxPos;
             if (searchRadius >= 0)
             {
-                int r = Mathf.FloorToInt(Mathf.Min(searchRadius, (end - start) / 2f));
-                start = startSegment - r;
-                end = startSegment + r + 1;
-                if (!Looped)
+                if (Looped)
                 {
-                    start = Mathf.Max(start, MinPos);
-                    end = Mathf.Min(end, MaxPos);
+                    var r = Mathf.Min(searchRadius, Mathf.FloorToInt((end - start) / 2f));
+                    start = startSegment - r;
+                    end = startSegment + r + 1;
+                }
+                else
+                {
+                    start = Mathf.Max(startSegment - searchRadius, MinPos);
+                    end = Mathf.Min(startSegment + searchRadius + 1, MaxPos);
                 }
             }
             stepsPerSegment = Mathf.RoundToInt(Mathf.Clamp(stepsPerSegment, 1f, 100f));