+++ com.unity.cinemachine/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Optimized path gizmo drawing.  Now 3-5 times faster.
 - TargetGroup now ignores members whose gameObjects are inactive.
 - CinemachinePathBase search radius fixed for not looped paths.
+- Fix priority comparison overflow.
 
 
 ## [2.9.4] - 2022-11-18