+++ com.unity.cinemachine/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Optimized path gizmo drawing.  Now 3-5 times faster.
 - TargetGroup now ignores members whose gameObjects are inactive.
 - CinemachinePathBase search radius fixed for not looped paths.
-- Bugfix: priority ordering was wrong when priority values were close to the integer min and max values.
+- Bugfix: priority ordering was wrong when the difference between any priority values were smaller than integer min or bigger than integer max values.
 
 
 ## [2.9.4] - 2022-11-18